If event is active you are not able to delete it.

+++ server/src/main/java/com/billbasher/services/EventService.java
@@ -2,6 +2,9 @@ package com.billbasher.services;
 
 import com.billbasher.model.EventDAO;
 import com.billbasher.repository.EventRep;
+import com.billbasher.repository.UserEventRep;
+
+import com.billbasher.repository.UserRep;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,6 +21,12 @@ public class EventService {
 
     @Autowired
     private ExpenseService expenseService;
+    @Autowired
+    private UserEventRep userEventRepository;
+    @Autowired
+    private UserRep userRepository;
+
+
 
     public EventDAO findEventById(@PathVariable("id") Long id) {
         return eventRepository.findById(id).get();
@@ -25,19 +34,33 @@ public class EventService {
     }
 
     public EventDAO updateEventById(Long id, EventDAO event) {
+        EventDAO existingEvent = findEventById(id);
+        if (existingEvent != null) {
+            // Update the existing event with the new data
+            existingEvent.setEventName(event.getEventName());
 
-        return eventRepository.save(event);
+            return eventRepository.save(existingEvent);
+        }
+        return null;
     }
 
     public void deleteEventById(Long id) {
         EventDAO event = findEventById(id);
         if (event != null) {
-            // Delete all expenses associated with the event
-            expenseService.deleteExpensesByEvent(event);
-            // Then delete the event
-            eventRepository.delete(event);
+            // Check if the event is finished
+            if (!event.getEventActive()) {
+                // Delete all expenses associated with the event
+                expenseService.deleteExpensesByEvent(event);
+                // Then delete the event
+                eventRepository.delete(event);
+            } else {
+                // Event cannot be deleted
+                throw new IllegalStateException("Event cannot be deleted as it is not finished or has more than one user.");
+            }
         }
     }
+
+
 
     public List<EventDAO> getAllEvents() {
 
@@ -48,4 +71,5 @@ public class EventService {
 
         return eventRepository.save(event);
     }
+
 }
